Добавил ввод и вывод

+++ mp2-lab1-set-master/include/tbitfield.h
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-typedef unsigned short int TELEM;
+typedef unsigned int TELEM;
 
 class TBitField
 {
